Merge PyUni::IO namespace into PyUni namespace.
We do not need a separate namespace here. The reason for the namespacing was in
tt3d, where we had a strict module layout. This is just a wrapper library for
the python client. Our current namspace policy is, to put everything in a name-
space which will (or may) get a separate python package. Everything else is
used all over the cpphost program and its just bloat to subnamespace it.

(thus pending: Put windowing in a namespace, but we'll wait until the merge
from cpp-port-sebi)