Implement log stream support
This allows a user to call LogBase::log(Severity) and get a reference to
an object which supports the std::ostream interface. Thus one can pipe
contents to the logs like::

    log->log(Error) << "foobar" << 23 << submit;

where submit is also defined in Log.hpp. Note that piping submit into
the LogPipe will destroy that instance (which is a bit of a hack but
neccessary).